Person Controller | Service

+++ src/main/java/com/accounts/repo/PersonRepo.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface PersonRepo extends JpaRepository<Person, Integer> {
+public interface PersonRepo extends JpaRepository<Person, Long> {
+
 }
